requirements.txt
# EEG Epilepsy Classification System
# Core dependencies

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0

# Visualization
matplotlib>=3.4.0

# Dimensionality reduction (optional)
umap-learn>=0.5.0

# For comparison/validation only (not used in implementation)
PyWavelets>=1.1.0

